+++ src/interfaces/requests/users/create_user_request.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize)]
 pub struct CreateUserReq {
-    pub user_id: u32,
+    pub user_id: u64,
     pub first_name: String,
     pub last_name: String,
     pub email: Option<String>,
